+++ NEWS.md
@@ -4,6 +4,8 @@
 
 -   Added the `mln.mean.sd()` function, which implements the Method for
     Unknown Non-Normal Distributions (MLN) approach of Cai et al. (2021)
+-   Added parametric bootstrap SE estimators for the QE, BC, and MLN
+    approaches. See the `get_SE()` generic function.
 
 ### Changes in Version 0.2.1 (2020-08-12)